Refactor AvdManager to store ILogger as a field
All callers (AvdManagerConnection and AvdManagerCli) pass the same
ILogger instance in all cases. Thus, rather than passing this parameter
everywhere, store it as a field. This also allows simplifying the
DeviceManager field.

Store the cached valid AVDs as an ImmutableList, rather than an
array (which is vulnerable to mutation by the callers).

Bug: n/a
Test: existing
Change-Id: I3be045e81930f61afc37ebb020580a3c74383810