Reset latestForegroundProcess when stop tracking
This change resets the latest seen foreground process when tracking is
stopped. This is necessary to cover the case when foreground process
tracking is stopped and resumed on the same device, while the
foreground process on the device has not changed.

If we don't resent the last seen foreground process, when tracking is
resumed it will assume that the process hasn't changed and will not send
a foreground process event to Studio.

Test: foreground_process_tracker_test.cc
Bug: 232067042
Change-Id: Id6231e1de84683bcc05a755d7afa6036b1c3c732