adblib: Add `deviceProperties()` API
This change moves `getprop` from the adblib-tools library
to the adblib "core" library. We will need this new `deviceProperties()`
service to be able to retrieve the API level of any device.

Retrieving the API level of a device is a low level enough capability
that is makes more sense to move to the "core' adblib library.

This change also introduces a `toImmutableMap` utility function to
prevent cache map from being modified by consumers.

This change also modifies `FakeAdbProvider` to store the last
created `TestAdbChannelProvider` instance to make testing easier.

Bug: 188067646
Test: Existing
Change-Id: I8521e43f573cab8c756c7eee52a4b76933f929ed