adblib: Wait for device before tracking it
`AdbLibDeviceClientManager` instances are created as soon as ddmlib
sees a new connected device. Since adblib device tracker run on a
different thread and is asynchronous, we can sometimes get notified
of a new device before it is known to the adblib device tracker.

The workaround is to wait for a couple of seconds max. for a device
to show up in the adblib tracker before tracking processes on
the newly connected device.

Bug: 188067646
Test: Included
Change-Id: I1a779545a04afe819797738453b0e108f14f576c