Update ISO 639-1, 639-2 and ISO 3166-2 codes
This adds around 5 regions and removes around 40 languages
(from 531 to 490). I've looked these up and they appear to
be from ISO 639-3, which AFAICT isn't supported yet in Android.
I added some code here to merge in 639-3, and that added over
7000 languages (so I disabled that for now). I'm guessing
last time I ran the updater many years ago these must have
been present in the ICU data or a 639-2 draft.

The default region for language code is obsolete now; the
timezone data is still updated here but in a followup CL
I will remove all of it.

Test: LocaleManagerTest and others
Fixes: 236442552
Change-Id: I659ba6815139d690172fa6e9c421a603c13c55ce