Report incidents on overloaded, indexed setter =, not array access [
When KTIJ-17726 is bundled, for overloaded, indexed setter, we will see
duplicate incidents on binary operator (=) and array access ([).
It is intuitive to report such incidents on only =.

This commit brings (slightly revised version of) KTIJ-17726 as an
extension function for UBinaryExpression.resolveOperator, then redirect
usages of existing resolveOperator to that extension.

At the same time, UArrayAccessExpression.resolveOperator will drop the
resolution result to the overloaded, indexed setter, if any, since the
aforementioned extension will handle that.

BinaryExpressionAsCallExpression is updated to retrieve receiver and
value arguments accordingly, if it models overloaded, indexed setter.

Exit plan: when UAST w/ KTIJ-17726 is coming, we can simply remove
UBinaryExpression.resolveOperatorExtended, since the origial
resolveOperator will do the same, and revert redirect resolveOperator
calls.

Bug: 238466859
Test: existing/updated tests
Change-Id: If608be5ab2a185f9f361acc5fb2380420e994e24