Machine-portable C/C++ path hashes
This change is a result of a meeting with androidx and reclient
(go/reclient) teams. The fact that hash paths change on a per-machine
basis is problematic because it would limit cache sharing.

This change makes the hashcode only depend on relative paths within
the project (along with other CMake/ndk-build flags).

I added a test that shows the path is stable between my local machine
and Linux, Mac, and Windows build-bots. My understanding is that
reclient uses Docker container linux build even for Windows and Mac.

Bug: NA
Test: Added a test

Change-Id: I132550e517a4d0241002af45b70c49a45c4d70da