Remove a spammy warning
In unit tests for which the threading agent is loaded (which can be done using the test_agents argument of a iml_module bazel rule) these warnings will be logged by all the tests that execute the threading annotated code but that don't actually test for the thread correctness (e.g. by installing a ThreadingCheckRule).

Bug: 228841408
Test: Unit tests
Change-Id: Ieccf1322d99a5800b0f5def5142fb756859ba8d9