Fix issues with ProviderBasedDirectoryEntryImpl
The class was calling ConfigurableFileTree.from with a provider of a set
of FileSystemLocation, while the function expects a single directory
element. This caused Gradle to throw UnsupportedNotationException as it
tries to parse the set as a single file. The problem impacted only flat
sources, so it wasn't captured by the jni tests.

This change maps the set into a single directory provider to be able to
use the configurable file tree API correctly. This problem has caused
the androidx integration branch to fail when adding a new API for java
resources as they were adding a java resource in an afterEvaluate block.

This change also adds the sourceset filteration that was not used before.

Test: Extended SourceSetsMixedApiUseTest with an assert that failed previously
Bug: N/A
Change-Id: I687eab1da6d3b4a8549de428e5a28a1fac2b66db